% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{assign_list}
\alias{assign_list}
\title{Assign values from a named list into an environment}
\usage{
assign_list(a_list, pos = -1, envir = as.environment(pos), inherits = FALSE)
}
\arguments{
\item{a_list}{Named list whose values will be assigned into the target
environment.}

\item{pos}{Position of the target environment on the search path.}

\item{envir}{Environment in which to assign the objects.}

\item{inherits}{Passed through to [base::assign()].}
}
\value{
Invisibly returns `NULL`.
}
\description{
Utility function based on `assignList()` from the package `tis` by Jeff
Hallman (https://github.com/cran/tis/blob/master/R/assignList.R).
Assigns the values in a list to variables in an environment. The variable names are taken from the names of the list, so all of the elements of the list must have non-blank names.
}
