% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{assignList}
\alias{assignList}
\title{Assign values from a named list into an environment}
\usage{
assignList(aList, pos = -1, envir = as.environment(pos), inherits = FALSE)
}
\arguments{
\item{aList}{Deprecated; use `a_list`.}

\item{pos}{Position of the target environment on the search path.}

\item{envir}{Environment in which to assign the objects.}

\item{inherits}{Passed through to [base::assign()].}
}
\value{
Invisibly returns `NULL`.
}
\description{
Deprecated compatibility wrapper for [assign_list()].
}
