% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{check_platform_ok}
\alias{check_platform_ok}
\title{Check whether the platform is Windows and the filename contains "http": if so, return a warning}
\usage{
check_platform_ok(input_stem)
}
\arguments{
\item{input_stem}{A netcdf filename or file stem}
}
\value{
TRUE or stop error
}
\description{
Utility function for the eReefs package
}
