% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{daysIn}
\alias{daysIn}
\title{Calculate the number of days in the month of the specified date}
\usage{
daysIn(d)
}
\arguments{
\item{d}{A date in the format given by as.Date}
}
\value{
An integer
}
\description{
With thanks to G.Grothendieck on StackOverflow.
}
