% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_discharge.R
\name{download_discharge}
\alias{download_discharge}
\title{Download river discharge (flow) data from Qld. Govt. servers at https://water-monitoring.information.qld.gov.au}
\usage{
download_discharge(
  station = "124001B",
  start_time = "20101001",
  end_time = "20180930",
  interval = "year"
)
}
\arguments{
\item{station}{Guaging station ID (obtainable from https://water-monitoring.information.qld.gov.au). Default is "124001B".
Stations used in eReefs models are:
        oconnell: "124001B")
        tully: "113006A"
        pioneer: "125016A"
        burnett: "136007A"
        fitzroy: "130005A"
        herbert: "116001F"
        normanby: "105107A"
        burdekin: "120006B"
        mulgrave: "111007A"
        russell: "111101D"
        barron: "110001D"
        daintree: "108002A"
        sjohnstone: "112101B"
        njohnstone: "112004A"}

\item{start_time}{Date from which to extract data, as text "YYYYMMDD". Default is "20101001"}

\item{end_time}{Last date required, as text "YYYYMMDD". Default is "20180930"}

\item{interval}{Time interval over which to average discharge, e.g. "year", "week", "day". Default is "year".}
}
\value{
mean discharge in ML/d
}
\description{
Download river discharge (flow) data from Qld. Govt. servers at https://water-monitoring.information.qld.gov.au
}
