% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R,
%   R/vertical_profiles.R
\name{earth.dist}
\alias{earth.dist}
\title{Calculate rough distance in kilometers between two points}
\usage{
earth.dist(long1, lat1, long2, lat2)

earth.dist(long1, lat1, long2, lat2)
}
\arguments{
\item{long1}{Longitude of the first point, in decimal degrees.}

\item{lat1}{Latitude of the first point, in decimal degrees.}

\item{long2}{Longitude of the second point, in decimal degrees.}

\item{lat2}{Latitude of the second point, in decimal degrees.}
}
\value{
Approximate distance in kilometres.
}
\description{
Not exported. This is very approximate - a package is available if a more accurate distance is needed.

Not exported. This is very approximate - a package is available if a more accurate distance is needed.
}
