% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{get_chron_date}
\alias{get_chron_date}
\title{A simple function to convert a date provided in any of several formats to a chron date}
\usage{
get_chron_date(d)
}
\arguments{
\item{d}{The date of interest. Can be any of:
c(year, month, day)
c(year, month, day, hour) 
Date format date (e.g., as.Date('1970-01-01', origin='1970-01-01'))
chron date
character format, e.g. '1970-01-01'}
}
\value{
date in chron format
}
\description{
A simple function to convert a date provided in any of several formats to a chron date
}
