% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{get_date_time}
\alias{get_date_time}
\title{Convert supported date formats to a `POSIXct` date-time}
\usage{
get_date_time(d)
}
\arguments{
\item{d}{The date of interest. Supported formats include:
`c(year, month, day)`
`c(year, month, day, hour)`
`c(year, month, day, hour, minute)`
`c(year, month, day, hour, minute, second)`
`YYYYMMDD`
`YYYYMMDDHHMM`
`YYYYMMDDHHMMSS`
`Date`
`POSIXct`
character format, e.g. `"1970-01-01"` or `"1970-01-01 12:00:00"`}
}
\value{
A `POSIXct` date-time value.
}
\description{
Unless a `POSIXct` value is supplied directly, parsed dates are interpreted
in the package's default timezone, `Etc/GMT-10`. Date-only inputs default to
midday to reduce accidental snapping to the previous or next model output.
}
