% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_profiles.R
\name{get_ereefs_vertical_slice}
\alias{get_ereefs_vertical_slice}
\title{Extract vertical profiles from an array of specified latitudes and longitudes over a specified time-period from an eReefs or other EMS netcdf file.}
\usage{
get_ereefs_vertical_slice(var_names = c("Chl_a_sum", "TN"),
  location_latlon = data.frame(latitude = c(-20, -16), longitude = c(148.5,
  152)), target_date = c(2016, 2, 4),
  input_file = "http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_bgc_GBR4_H2p0_B2p0_Chyd_Dcrt/gbr4_bgc_simple_2010-01.nc",
  input_grid = NA, eta_stem = NA)
}
\arguments{
\item{location_latlon}{A data frame of latitudes and longitudes.  Defaults to data.frame(latitude=c(-20, -16), longitude=c(148.5, 152.0)).
If length(location_lat_lon)==2, extract every grid cell along a straight line between the two points specified.
Otherwise, extract only the locations corresponding to the cells nearest the specified points.}

\item{target_date}{Target date to extract profile. Can be a date, or text formatted for as.Date(), or a (year, month, day) vector.
Defaults to c(2016, 02, 04).}

\item{input_file}{is the URI or file location of any of the EMS output files, 
Defaults to "http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_bgc_GBR4_H2p0_B2p0_Chyd_Dcrt/gbr4_bgc_simple_2010-01.nc". 
If using Windows, you will need to set this to a local inputfile stem.}

\item{input_grid}{Name of the locally-stored or opendap-served netcdf file that contains the grid
coordinates for the top and bottom of each layer (z_grid). If not specified, the function will first look for
z_grid can be found in the first INPUT_STEM file, and if not found, will check whether the size 
of the variables in the input file corresponds to the size expected for GBR4 or GBR1, and load an appropriate 
z grid from data files stored in this package. Alternatively, you can provide the location of a full 
(not simple-format) ereefs netcdf output file such as 
"http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_hydro_all/gbr4_all_2016-09.nc"}

\item{eta_stem}{The URI or file location of the model output files that contains the surface elevation (eta), minus the
date components of the filename in the case of GBR1 or GBR4 files, and ommitting the file extension, ".nc". Needed
only if eta is not in the file indicated by input_file (e.g. some GBR1 bgc files).}

\item{var_name}{A vector of EMS variable names. Defailts to c('Chl_a_sum', 'TN'))}
}
\value{
a list containing a vector of dates, an array of surface elevations (eta), the vertical grid (z_grid) and a data frame of values.
}
\description{
This is slow because we extract one profile at a time to avoid memory issues with large netcdf files.
}
