% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{get_file_stem}
\alias{get_file_stem}
\title{Returns the 'stem' of a filename (i.e. the filename with the date and extension cut off the end)}
\usage{
get_file_stem(filename)
}
\arguments{
\item{filename}{A netcdf filename from an EMS model run}
}
\value{
a character string
}
\description{
Utility function for ereefs package
}
