% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_profiles.R
\name{plot_ereefs_slice}
\alias{plot_ereefs_slice}
\title{Produces a coloured tile plot of a vertical slice already fetched from an eReefs or other EMS netcdf file.}
\usage{
plot_ereefs_slice(
  slice,
  var_name = "Chl_a_sum",
  scale_col = "viridis",
  scale_lim = NA,
  var_units = ""
)
}
\arguments{
\item{slice}{A list object as returned by [get_ereefs_slice()].}

\item{var_name}{Name of the variable to display from `slice$values`.}

\item{scale_col}{Colour scale specification. `"viridis"` is the default for
scalar slices.}

\item{scale_lim}{values for low and high limits of colourscale. Defaults to
the full range of the plotted data.}

\item{var_units}{Units label to show on the fill scale.}
}
\value{
p handle for the generated figure
}
\description{
Relies on output from get_ereefs_slice().
}
