% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plume_class}
\alias{plume_class}
\title{Classify optical plume type from modelled reflectances}
\usage{
plume_class(rsr)
}
\arguments{
\item{rsr}{List of 2D reflectance arrays in wavelength order:
`R_412`, `R_443`, `R_488`, `R_531`, `R_547`, `R_667`, `R_678`.}
}
\value{
An integer matrix of the same horizontal shape as the reflectance
  inputs, with plume class values between 1 and 7.
}
\description{
Calculates plume optical classes from eReefs/EMS reflectance outputs using
the Devlin et al. (2012) class definitions as adapted for this package.
}
