% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly2spatialpoly.R
\name{poly2sp}
\alias{poly2sp}
\title{Backward-compatible wrapper for older function name.}
\usage{
poly2sp(polydf)
}
\arguments{
\item{polydf}{Polygon data frame in the format used by `geom_polygon()`.}
}
\value{
A `terra::SpatVector`.
}
\description{
Backward-compatible wrapper for older function name.
}
