% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly2spatialpoly.R
\name{poly2sv}
\alias{poly2sv}
\title{Convert a polygon dataframe in the format used by ggplot2::geom_polygon to a SpatVector}
\usage{
poly2sv(polydf)
}
\arguments{
\item{polydf}{A polygon in the format required by geom_polygon: a data.frame with values, x and y positions and id labels}
}
\value{
A SpatVector as used in the package 'terra'
}
\description{
Assumes that the dataframe only contains rombuses.
}
