% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly2spatialpoly.R
\name{sp2raster}
\alias{sp2raster}
\title{Backward-compatible wrapper for older function name.}
\usage{
sp2raster(
  sv,
  xmn = 142.45,
  ymn = -27.5,
  resolution = 0.01,
  xmx = NA,
  ymx = NA,
  r = NULL
)
}
\arguments{
\item{sv}{SpatVector object, e.g. as output by poly2sv()}

\item{xmn}{Minimum x value to use for the grid. Defaults to 142.45}

\item{ymn}{Minimum y value for the grid. Defaults to -27.5}

\item{resolution}{Grid resolution in degrees. Defaults to 0.01}

\item{xmx}{Maximum x value for the grid. Defaults to the maximum x extent of
`sv`.}

\item{ymx}{Maximum y value for the grid. Defaults to the maximum y extent of
`sv`.}

\item{r}{A raster object with the correct grid set up (optional). Default=NULL}
}
\value{
A `terra` raster.
}
\description{
Backward-compatible wrapper for older function name.
}
