% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{substitute_filename}
\alias{substitute_filename}
\title{Check whether the given filename is a shortcut and if so, set the full filename}
\usage{
substitute_filename(input_file)
}
\arguments{
\item{input_file}{A netcdf filename or file stem}
}
\value{
input_file
}
\description{
Utility function for the eReefs package
}
