% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_profiles.R
\name{earth.dist}
\alias{earth.dist}
\alias{earth.dist}
\title{Calculate rough distance in kilometers between two points}
\usage{
earth.dist(long1, lat1, long2, lat2)

earth.dist(long1, lat1, long2, lat2)
}
\description{
Not exported. This is very approximate - a package is available if a more accurate distance is needed.

Not exported. This is very approximate - a package is available if a more accurate distance is needed.
}
