% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ereefs-package.R
\docType{package}
\name{ereefs-package}
\alias{ereefs-package}
\alias{ereefs}
\title{ereefs package}
\description{
A package for accessing, extracting, and visualising eReefs and
  other CSIRO-EMS model output. The current development workflow is designed
  around local NetCDF files, live OPeNDAP datasets, and THREDDS catalog
  services, with support for both curvilinear EMS grids and centre-only
  regular grids.
}
\references{
https://www.ereefs.org.au/about/
https://research.csiro.au/ereefs/welcome-to-ereefs/about-ereefs/
https://ereefs.aims.gov.au/ereefs-aims
https://www.barrierreef.org/what-we-do/projects/eReefs
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ereefs.info}
}

}
\author{
\strong{Maintainer}: Barbara Robson \email{B.Robson@aims.gov.au}

Other contributors:
\itemize{
  \item Mark Baird \email{Mark.Baird@csiro.au} [contributor]
  \item Diego Barneche \email{d.barneche@aims.gov.au} [contributor]
}

}
\keyword{internal}
