% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_profiles.R
\name{find_intersections}
\alias{find_intersections}
\title{An internal utility function to find the grid line intersections of a line segment}
\usage{
find_intersections(
  location_latlon,
  x_grid,
  y_grid,
  latitude,
  longitude,
  first_point = FALSE
)
}
\description{
An internal utility function to find the grid line intersections of a line segment
}
