% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{get_date_time}
\alias{get_date_time}
\title{A simple function to convert a date provided in any of several formats to a date-time object.
Assumes GMT-10.}
\usage{
get_date_time(d)
}
\arguments{
\item{d}{The date of interest. Can be any of:
c(yr, month, day)
c(yr, month, day, hour) 
c(yr, month, day, hour, minute)
c(yr, month, day, hour, minute, second) 
YYYYMMDD
YYYYMMDDHHMM
YYYYMMDDHHMMSS
Date format date (e.g., as.Date('1970-01-01', origin='1970-01-01'))
character format, e.g. '1970-01-01'}
}
\value{
date in date-time format (for lubridate)
}
\description{
A simple function to convert a date provided in any of several formats to a date-time object.
Assumes GMT-10.
}
