% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{get_ereefs_bottom_ts}
\alias{get_ereefs_bottom_ts}
\title{Extracts time-series of selected variables from the bottom water-column cell at specified locations from eReefs output files}
\usage{
get_ereefs_bottom_ts(var_names = c("Chl_a_sum", "TN"),
  location_latlon = c(-23.39189, 150.88852), start_date = c(2010, 12,
  31), end_date = c(2016, 12, 31), input_file = "menu",
  input_grid = NA, eta_stem = NA, override_positive = FALSE,
  verbosity = 1)
}
\arguments{
\item{var_names}{either a single character value or a vector specifying the short names for variables that you 
want from the netcdf file. Defaults to c('Chl_a_sum', 'TN').}

\item{location_latlon}{is a vector containing the decimal latitude and longitude of the desired location. If 
you want to specify an x-y grid coordinate instead of a latitude and longitude, you can: to do this, 
is.integer(location_latlon) must be TRUE. Defaults to c(-23.39189, 150.88852).}

\item{start_date}{date  for animation. Can either be a) a vector of integers in the format 
c(year, month, day); b) a date obtained e.g. from as.Date(); or c) a character string 
     formatted for input to as.Date(). Defaults to c(2010,12,31).}

\item{input_file}{is the URI or file location of any of the EMS output files, 
Defaults to a menu selection. Set to "choices" to see some other pre-defined options that
can be used (codenames as used in https://research.csiro.au/ereefs/models/model-outputs/access-to-raw-model-output/ )}

\item{input_grid}{Name of the locally-stored or opendap-served netcdf file that contains the grid
coordinates for the top and bottom of each layer (z_grid). If not specified, the function will first look for
z_grid can be found in the first INPUT_STEM file, and if not found, will check whether the size 
of the variables in the input file corresponds to the size expected for GBR4 or GBR1, and load an appropriate 
z grid from data files stored in this package. Alternatively, you can provide the location of a full 
(not simple-format) ereefs netcdf output file such as 
"http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_hydro_all/gbr4_all_2016-09.nc"
"http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_hydro_all/gbr4_all_2016-09.nc".}

\item{eta_stem}{The URI or file location of the model output files that contains the surface elevation (eta), minus the
date components of the filename in the case of GBR1 or GBR4 files, and ommitting the file extension, ".nc". Needed
only if eta is not in the files indicated by input_stem (e.g. some GBR1 bgc files).}

\item{override_positive}{Reverse the value of the "positive" attribute of botz for BGC files, assuming that it is
incorrect. Default FALSE}

\item{end}{date for animation. Can either be a) a vector of integers in the format 
c(year, month, day); b) a date obtained e.g. from as.Date(); or c) a character string 
     formatted for input to as.Date(). Defaults to c(2016,12,31).}
}
\value{
a data frame containing the dates and values of extracted variables
}
\description{
See also get_ereefs_ts() to extract from a specified layer instead of a depth-integrated value.
By Barbara Robson (AIMS).
}
\examples{
\dontrun{
get_ereefs_bottom_ts(c('Chl_a_sum', 'NH4'), location_latlon=data.frame(latitide=-23.39189, longitude=150.88852), layer='surface', start_date=c(2010,12,31),end_date=c(2011,1,5), input_file=2)
}
}
