% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{get_ereefs_depth_integrated_ts}
\alias{get_ereefs_depth_integrated_ts}
\title{Extracts depth-integrated time-series of selected variables at specified locations from eReefs output files}
\usage{
get_ereefs_depth_integrated_ts(
  var_names = c("Chl_a_sum", "TN"),
  geocoordinates = c(-23.39189, 150.88852),
  start_date = c(2016, 2, 4),
  end_date = c(2016, 3, 2),
  input_file = "catalog",
  input_grid = NA,
  eta_stem = NA,
  override_positive = FALSE,
  verbosity = 1,
  mass = FALSE
)
}
\arguments{
\item{var_names}{either a single character value or a vector specifying the short names for variables that you 
want from the netcdf file. Defaults to c('Chl_a_sum', 'TN').}

\item{geocoordinates}{is a vector containing the decimal latitude and longitude of the desired location. If 
you want to specify an x-y grid coordinate instead of a latitude and longitude, you can: to do this, 
is.integer(geocoordinates) must be TRUE. Defaults to c(-23.39189, 150.88852).}

\item{start_date}{Start of the extraction period. Can be a `Date`,
`POSIXct`, character string, or a vector such as `c(year, month, day)`.
Date-only inputs default to midday in `Etc/GMT-10`.}

\item{end_date}{End of the extraction period, specified as for `start_date`.}

\item{input_file}{is the URL or file location of any of the EMS output files or a THREDDS catalog URI. 
Defaults to a menu selection based on current NCI catalogs. Can also be set to "nci", "menu" or "catalog" for the same behaviour.
Set to "old_menu" to provide old menu options instead of menu options from the NCI catalog.
Numeric values are interpreted as references to selections available from the old menu.
Short codes can be used for some options (codenames as used in https://research.csiro.au/ereefs/models/model-outputs/access-to-raw-model-output/ )}

\item{input_grid}{Name of the locally-stored or opendap-served netcdf file that contains the grid
coordinates for the top and bottom of each layer (z_grid). If not specified, the function will first look for
z_grid can be found in the first INPUT_STEM file, and if not found, will check whether the size 
of the variables in the input file corresponds to the size expected for GBR4 or GBR1, and load an appropriate 
z grid from data files stored in this package. Alternatively, you can provide the location of a full 
(not simple-format) ereefs netcdf output file such as 
"https://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_hydro_all/gbr4_all_2016-09.nc"
"https://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_hydro_all/gbr4_all_2016-09.nc".}

\item{eta_stem}{The URI or file location of the model output files that contains the surface elevation (eta), minus the
date components of the filename in the case of GBR1 or GBR4 files, and ommitting the file extension, ".nc". Needed
only if eta is not in the files indicated by input_stem (e.g. some GBR1 bgc files).}

\item{override_positive}{Reverse the value of the "positive" attribute of botz for BGC files, assuming that it is
incorrect. Default FALSE}

\item{verbosity}{(Defailt 1) how much do you want to know about progress?}

\item{mass}{(Default FALSE) Set to true if you want the mass per square metre rather than the mean concentration over depth returned}
}
\value{
a tibble containing the dates and values of extracted variables
}
\description{
See also get_ereefs_ts() to extract from a specified layer instead of a depth-integrated value.
By Barbara Robson (AIMS).
}
\examples{
\dontrun{
get_ereefs_depth_integrated_ts(c('Chl_a_sum', 'NH4'), geocoordinates=tibble(latitude=-23.39189, longitude=150.88852), start_date=c(2010,12,31),end_date=c(2011,1,5), input_file=2)
get_ereefs_depth_integrated_ts(c('salt', 'temp'), geocoordinates=tibble(latitude=c(-23.39189, -23.3), longitude=c(150.88852, 150.8)), start_date=c(2011,1,2),end_date=c(2011,1,5))
}
}
