% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{get_ereefs_depth_specified_ts}
\alias{get_ereefs_depth_specified_ts}
\title{Extracts time-series of selected variables from eReefs output files at a specified location and depth below the
surface.}
\usage{
get_ereefs_depth_specified_ts(var_names = c("Chl_a_sum", "TN"),
  location_latlon = c(-23.39189, 150.88852), depth = 1,
  start_date = c(2010, 12, 31), end_date = c(2016, 12, 31),
  input_stem = "http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_bgc_GBR4_H2p0_B2p0_Chyd_Dcrt/gbr4_bgc_simple_",
  grid_file = "http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_hydro_all/gbr4_all_2016-09.nc")
}
\arguments{
\item{var_names}{either a single character value or a vector specifying the short names for variables that you 
want from the netcdf file. Defaults to c('Chl_a_sum', 'TN').}

\item{location_latlon}{A vector containing the decimal latitude and longitude of the desired location. If 
you want to specify an x-y grid coordinate instead of a latitude and longitude, you can: to do this, 
is.integer(location_latlon) must be TRUE. Defaults to c(-23.39189, 150.88852).}

\item{depth}{Depth in metres below the surface. Default 1.0. If the bottom of the water is shallower than the specified depth,
return values from the bottom of the water column.}

\item{start_date}{date  for animation. Can either be a) a vector of integers in the format 
c(year, month, day); b) a date obtained e.g. from as.Date(); or c) a character string 
     formatted for input to as.Date(). Defaults to c(2010,12,31).}

\item{input_stem}{is the URI or file location of the eReefs output files, ommitting the year and month parts of the filenames and ommitting ".nc" . Defaults to "http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_bgc_GBR4_H2p0_B2p0_Chyd_Dcrt/gbr4_bgc_simple_". If using Windows, you will need to set this to a local filename stem.}

\item{grid_file}{is the location of the netcdf file that contains the vertical grid information -- likely an output file
from the hydrodynamic model used to drive the BGC model. Defaults to"http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_hydro_all/gbr4_all_2016-09.nc"}

\item{end}{date for animation. Can either be a) a vector of integers in the format 
c(year, month, day); b) a date obtained e.g. from as.Date(); or c) a character string 
     formatted for input to as.Date(). Defaults to c(2016,12,31).}
}
\value{
a data frame containing the dates and values of extracted variables
}
\description{
Note that this function can use an OpenDAP URI if you are running it under Linux or MacOS, but 
not (as at February 2018) under Windows, because of issues with the windows version of the netcdf4 library. 
Under Windows, you can still use it for locally-saved files. See also get_ereefs_ts() to extract from a specified 
layer instead of a depth-integrated value and get_ereefs_z_integrated_ts() to calculate depth-integrated values.
By Barbara Robson (AIMS).
}
\examples{
get_ereefs_z_integrated_ts('Chl_a_sum', c(-23.39189, 150.88852), layer='surface')
}
