% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{get_ereefs_ts}
\alias{get_ereefs_ts}
\title{Extracts time series at a specified location from eReefs model output files}
\usage{
get_ereefs_ts(var_names = c("Chl_a_sum", "TN"),
  location_latlon = c(-23.39189, 150.88852), layer = "surface",
  start_date = c(2010, 12, 31), end_date = c(2016, 10, 31),
  input_file = "http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_bgc_GBR4_H2p0_B2p0_Chyd_Dcrt/gbr4_bgc_simple_2010-01.nc")
}
\arguments{
\item{var_names}{either a single character value or a vector specifying the short names for variables that you 
want from the netcdf file. Defaults to c('Chl_a_sum', 'TN').}

\item{location_latlon}{is a vector containing the decimal latitude and longitude of the desired location. If 
you want to specify an x-y grid coordinate instead of a latitude and longitude, you can: to do this, 
is.integer(location_latlon) must be TRUE. Defaults to c(-23.39189, 150.88852).}

\item{layer}{is the vertical grid layer to extract, or 'surface' to get the surface value. Defaults to 'surface'.}

\item{start_date}{date  for animation. Can either be a) a vector of integers in the format 
c(year, month, day); b) a date obtained e.g. from as.Date(); or c) a character string 
     formatted for input to as.Date(). Defaults to c(2010,12,31).}

\item{input_file}{is the URI or file location of any of the EMS output files, 
Defaults to "http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_bgc_GBR4_H2p0_B2p0_Chyd_Dcrt/gbr4_bgc_simple_2010-01.nc". 
If using Windows, you will need to set this to a local inputfile stem.}

\item{end}{date for animation. Can either be a) a vector of integers in the format 
c(year, month, day); b) a date obtained e.g. from as.Date(); or c) a character string 
     formatted for input to as.Date(). Defaults to c(2016,10,31).}
}
\value{
a data frame containing the dates and values of extracted variables.
}
\description{
Create a time-series of values of one or more selected model output variables in a specified layer of the
water column or sediment store (by default, the surface layer), aat a specified geographic location within 
the model domain.  See also get_ereefs_depth_integrated_ts() to extract depth-integrated values and 
get_ereefs_depth_specified_ts() to extract values at a specified depth below the free (tidally moving) 
surface. Note that this function can use an OpenDAP URI if you are running it under Linux or MacOS, but 
not (as at February 2018) under Windows, because of issues with the windows version of the netcdf4 library. 
Under Windows, you can still use it for locally-saved files. Barbara Robson (AIMS).
}
\examples{
get_ereefs_ts('Chl_a_sum', c(-23.39189, 150.88852), layer='surface')
}
