% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{get_origin_and_times}
\alias{get_origin_and_times}
\title{Opens the netcdf file, input_file, and extracts the origin (reference date/time) and time dimension.}
\usage{
get_origin_and_times(input_file, as_chron = "TRUE")
}
\arguments{
\item{input_file}{The name of the netcdf file (in standard or simple EMS netcdf format) from which
to extract the data}
}
\value{
A list containing ereefs_origin (the reference data/time used in the netcdf file) and the time series, ds
}
\description{
Opens the netcdf file, input_file, and extracts the origin (reference date/time) and time dimension.
}
