% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{get_params}
\alias{get_params}
\title{Set up various parameters needed by many of the data extraction and plotting functions}
\usage{
get_params(start_date, end_date, input_file, var_names)
}
\value{
list of parameter values
}
\description{
Set up various parameters needed by many of the data extraction and plotting functions
}
