% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{map_ereefs}
\alias{map_ereefs}
\title{Create a surface map of eReefs model output.}
\usage{
map_ereefs(var_name = "true_colour", target_date = c(2011, 1, 30),
  layer = "surface", Google_map_underlay = TRUE,
  input_stem = "http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_bgc_GBR4_H2p0_B2p0_Chyd_Dcrt/gbr4_bgc_simple_",
  input_grid = "http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_hydro_all/gbr4_all_2016-09.nc",
  scale_col = c("ivory", "coral4"), zoom = 6)
}
\arguments{
\item{var_name}{Short name of the variable to plot. This can be any variable in the
eReefs netcdf file that you are accessing (refer to eReefs model
documentation or extract variable names from the netcdf file for a full 
list). In addition, two special var_name values are supported: 'true_colour'
and 'plume'. 'true_colour' provides a simulated MODIS true colour
image (refer to Baird et al., 2016 for en explanation).'plume'
provides a map of calculated plume colour class as per Devlin et al. (2012)
and Baird et al. (2017). Defaults to true_colour.}

\item{target_date}{Date to map. Can either be a) a vector of integers in the format 
c(year, month, day); b) a date obtained e.g. from as.Date(); or c) a character string 
     formatted for input to as.Date(). Defaults to c(2011,1,30).}

\item{layer}{Either an integer layer number or 'surface' to choose the surface layer. Defaults to 'surface'.}

\item{Google_map_underlay}{Set to TRUE (the default) to use ggmap to show a Google Map as
an underlay for the model output plot. Requires the ggmap librray.}

\item{input_stem}{Path to the locally-stored or opendap-served netcdf file, including the stem
of the file name (i.e. the start of the filename without the date or file extension).
Defaults to "http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_bgc_GBR4_H2p0_B2p0_Chyd_Dcrt/gbr4_bgc_simple_".,}

\item{input_grid}{Name of the locally-stored or opendap-served netcdf file that contains the grid
coordinates for the cell corners (x_grid and y_grid). This is usually the eReefs hydrodynamic model
output file corresponding to the biogeochemical model output file being used. 
Defaults to "http://dapds00.nci.org.au/thredds/dodsC/fx3/gbr4_hydro_all/gbr4_all_2016-09.nc"}

\item{scale_col}{Vector of colours to use for low and high values in the colour scale. This can be a colour 
from the ggplot colour palette or a RGB hash code. Ignored for true_colour plots. 
Defaults to c('ivory', 'coral4').}

\item{zoom}{Value of zoom passed to ggmap(). Set to 5 if you want to show the entire extent 
of eReefs models. Defaults to 6. Higher values will zoom in further.}
}
\value{
a ggplot object
}
\description{
Creates a colour map showing concentrations of a specified eReefs model output variable at a specified
model layer (by default, the surface layer). The map is optionally (and by default) overlain on a Google 
Earth map of the region. The function should work under Linux or MacOS. The R netcdf4 library doesn't 
handle opendap served files under Windows, unfortunately, but can be used under Windows for locally-stored
model output files. By Barbara Robson (AIMS).
}
\details{
References:

Baird, M.E., Cherukuru, N., Jones, E., Margvelashvili, N., Mongin, M., Oubelkheir, K., 
  Ralph, P.J., Rizwi, F., Robson, B.J., Schroeder, T. and Skerratt, J., 2016. Remote-sensing 
  reflectance and true colour produced by a coupled hydrodynamic, optical, sediment, 
  biogeochemical model of the Great Barrier Reef, Australia: comparison with satellite data. 
  Environmental Modelling & Software, 78, pp.79-96.

Baird, M.E., Andrewartha, J., Herzfeld, M., Jones, E.M., Margvelashvili, N., Mongin, M., Rizwi, 
  F.,  Skerratt, J., Soja-Wozniak, M., Wild-Allen, K., Schroe der, T., Robson, B.J., da Silva, E. 
  and Devlin, M., 2017. River plumes of the Great Barrier Reef: freshwater , sediment and optical 
  footprints quantified by the eReefs modelling system  In Syme, G., Hatton MacDonald, D., Fulton, 
  B. and Piantadosi, J. (eds) MODSIM2017, 22nd International Congress on Modelling and Simulation. 
  Modelling and Simulation Society of Australia and New Zealand, December 2017, pp. 894–900

Devlin, M.J., McKinna, L.W., Álvarez-Romero, J.G., Petus, C., Abott, B., Harkness, P. and 
  Brodie, J., 2012. Mapping the pollutants in surface riverine flood plume waters in the Great 
  Barrier Reef, Australia. Marine pollution bulletin, 65(4-9), pp.224-235.
}
\examples{
map_ereefs()
map_ereefs('TN')
map_ereefs('plume', target_date=c(2011, 1, 30))
map_ereefs('Chl_a_sum', target_date='2016-07-15',, scale_col=c('ivory', 'green4'))
}
