% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_profiles.R
\name{plot_ereefs_slice}
\alias{plot_ereefs_slice}
\title{Produces a coloured tile plot of a vertical slice already fetched from an eReefs or other EMS netcdf file.}
\usage{
plot_ereefs_slice(
  slice,
  var_name = "Chl_a_sum",
  scale_col = "spectral",
  scale_lim = NA,
  var_units = ""
)
}
\arguments{
\item{slice}{A list object as output by get_ereefs_slice(), containing dates, eta, z_grid, botz,
a data frame of values and a data frame of latitudes and longitudes}

\item{scale_col}{Vector of colours to use for low and high values in the colour scale. This can be a colour 
from the ggplot colour palette or a RGB hash code, or "spectral". Ignored for true_colour plots. 
Example: c('ivory', 'coral4').
If one value is given (other than "spectral"), low colour is set to ivory and high colour to the value given.
If three values are given, uses scale_fill_gradient2 (spectrum from low to high through middle value).
Defaults to "spectral"}

\item{scale_lim}{values for low and high limits of colourscale. Defaults to full range.}
}
\value{
p handle for the generated figure
}
\description{
Relies on output from get_ereefs_slice().
}
