% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_profiles.R
\name{plot_ereefs_zvt}
\alias{plot_ereefs_zvt}
\title{Produces a coloured rect plot of a vertical profile over time}
\usage{
plot_ereefs_zvt(
  slice,
  var_name = "Chl_a_sum",
  scale_col = c("ivory", "hotpink"),
  scale_lim = NA
)
}
\arguments{
\item{slice}{A list object as output by get_ereefs_profile(), containing
dates, eta, z_grid, botz, and profile values.}

\item{var_name}{Name of the variable to display from `slice$profiles`. If
`slice$profiles` is a squeezed single-variable matrix, this is used only
for the legend title.}

\item{scale_col}{Colours to use for low and high values. Default c("ivory", "hotpink").}

\item{scale_lim}{values for low and high limits of colourscale. Defaults to full range.}
}
\value{
p handle for the generated figure
}
\description{
Relies on output from get_ereefs_profile().
}
