% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_profiles.R
\name{plot_ereefs_zvt}
\alias{plot_ereefs_zvt}
\title{Produces a coloured rect plot of a vertical profile over time}
\usage{
plot_ereefs_zvt(
  slice,
  var_name = "Chl_a_sum",
  scale_col = c("ivory", "hotpink"),
  scale_lim = NA
)
}
\arguments{
\item{scale_col}{Colours to use for low and high values. Default c("ivory", "hotpink").}

\item{scale_lim}{values for low and high limits of colourscale. Defaults to full range.}

\item{profileObj}{A list object as output by get_ereefs_profile(), containing dates, eta, z_grid, botz,
and a data frame of values.}
}
\value{
p handle for the generated figure
}
\description{
Relies on output from get_ereefs_profile().
}
