% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R,
%   R/plotting_functions_old.R
\name{plume_class}
\alias{plume_class}
\title{Calculate and return the plume optical class from eReefs model output..}
\usage{
plume_class(rsr)

plume_class(rsr)
}
\arguments{
\item{rsr}{A list of 2D vectors containing the reflectances at various wavelengths
     extracted from an EMS netcdf file:
rsr <- list(R_412, R_443, R_488, R_531, R_547, R_667, R_678)}
}
\value{
an array of the same size as R_412 containing values between 1 and 7 correspodning
        to optical plume classes.

an array of the same size as R_412 containing values between 1 and 7 correspodning
        to optical plume classes.
}
\description{
Uses reflectance at multiple wavelengths from the model output to calculate plume colour
classes as defined by Devlin et al. (2012). Adapted from the Matlab function plume_detect.m 
by Mark Baird (CSIRO). This version by Barbara Robson (AIMS).

Uses reflectance at multiple wavelengths from the model output to calculate plume colour
classes as defined by Devlin et al. (2012). Adapted from the Matlab function plume_detect.m 
by Mark Baird (CSIRO). This version by Barbara Robson (AIMS).
}
\details{
Devlin, M.J., McKinna, L.W., Álvarez-Romero, J.G., Petus, C., Abott, B., Harkness, P. and 
  Brodie, J., 2012. Mapping the pollutants in surface riverine flood plume waters in the Great 
  Barrier Reef, Australia. Marine pollution bulletin, 65(4-9), pp.224-235.

Devlin, M.J., McKinna, L.W., Álvarez-Romero, J.G., Petus, C., Abott, B., Harkness, P. and 
  Brodie, J., 2012. Mapping the pollutants in surface riverine flood plume waters in the Great 
  Barrier Reef, Australia. Marine pollution bulletin, 65(4-9), pp.224-235.
}
\examples{
\dontrun{
plume_class(rsr)
}
\dontrun{
plume_class(rsr)
}
}
