% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly2spatialpoly.R
\name{poly2sp}
\alias{poly2sp}
\title{Convert a polygon dataframe in the format used by ggplot2::geom_polygon to a spatialPolygonDataFrame}
\usage{
poly2sp(polydf)
}
\arguments{
\item{polydf}{A polygon in the format required by geom_polygon: a data.frame with values, x and y positions and id labels}
}
\value{
A spatialPolygonsDataFrame object, as used in the package 'sp'
}
\description{
Assumes that the dataframe only contains rombuses.
}
