% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{safe_nc_open}
\alias{safe_nc_open}
\title{A wrapper to ncdf4::nc_open() that will pause and try again up to 119 times
if at first it fails, to overcome temporary net access errors or DAP errors.}
\usage{
safe_nc_open(filename, tries = 4)
}
\arguments{
\item{tries}{number of times to retry (increasing pause length by one second each time. Default 4}
}
\value{
variable extracted using ncvar_get()
}
\description{
Parameters before 'tries' are passed through to ncvar_get
}
