% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{safe_ncvar_get}
\alias{safe_ncvar_get}
\title{A wrapper to ncdf4::ncvar_get() that will pause and try again up to 119 times
if at first it fails, to overcome temporary net access errors or DAP errors.}
\usage{
safe_ncvar_get(
  nc,
  varid = NA,
  start = NA,
  count = NA,
  verbose = FALSE,
  signedbyte = TRUE,
  collapse_degen = TRUE,
  raw_datavals = FALSE,
  tries = 12
)
}
\arguments{
\item{tries}{number of times to retry (increasing pause length by one second each time. Default 12}
}
\value{
variable extracted using ncvar_get()
}
\description{
Parameters before 'tries' are passed through to ncvar_get
}
