% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly2spatialpoly.R
\name{sp2raster}
\alias{sp2raster}
\title{Convert a SpatialPolygonDataFrame to a raster. (Wrapper for raster::rasterize() with some reasonable defaults for the ereefs package)}
\usage{
sp2raster(sPdf, xmn = 142.45, ymn = -27.5, resolution = 0.01, xmx = NA,
  ymax = NA, r = NA)
}
\arguments{
\item{sPdf}{SpatialPolygonDataFrame object, e.g. as output by poly2sp.}

\item{xmn}{Minimum x value to use for the grid. Defaults to 142.45}

\item{ymn}{Minimum y value for the grid. Defaults to -27.5}

\item{resolution}{Grid resolution in degrees. Defaults to 0.01}

\item{ymax}{Maximum y value for the grid. Defaults to max value from the SpatialPolygonDataFrame bounding box.}

\item{xmax}{Maximum x value for the grid. Defaults to max value from the SpatialPolygonDataFrame bounding box.}
}
\value{
A SpatialPolygonsDataFrame object, as used in the package 'sp'
}
\description{
Convert a SpatialPolygonDataFrame to a raster. (Wrapper for raster::rasterize() with some reasonable defaults for the ereefs package)
}
