% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{substitute_filename}
\alias{substitute_filename}
\title{Check whether the given filename is a shortcut and if so, set the full filename}
\usage{
substitute_filename(input_file = "catalog")
}
\arguments{
\item{input_file}{is the URL or file location of any of the EMS output files or a THREDDS catalog URI. 
Defaults to a menu selection based on current NCI catalogs. Can also be set to "nci", "menu" or "catalog" for the same behaviour.
Set to "old_menu" to provide old menu options instead of menu options from the NCI catalog.
Numeric values are interpreted as references to selections available from the old menu.
Short codes can be used for some options (codenames as used in https://research.csiro.au/ereefs/models/model-outputs/access-to-raw-model-output/ )}
}
\value{
input_file
}
\description{
Utility function for the eReefs package
}
